model/requirements.txt
transformers
torch
pypdf2
python-docx 
pypandoc